rm 32b prover
