+++ prover/src/constraints/mod.rs
@@ -1,4 +1,7 @@
-pub mod constraints_templates;
-pub mod cpu_air;
-pub mod decode_air;
-pub mod utils;
+//! 64-bit VM prover constraint templates.
+//!
+//! This module provides constraint templates for the 64-bit VM prover
+//! using the Goldilocks field.
+
+pub mod cpu;
+pub mod templates;
