--- prover/src/constraints64/mod.rs
@@ -1,7 +1,0 @@
-//! 64-bit VM prover constraint templates.
-//!
-//! This module provides constraint templates for the 64-bit VM prover
-//! using the Goldilocks field.
-
-pub mod cpu;
-pub mod templates;
