+++ prover/src/lib.rs
@@ -1,6 +1,4 @@
 pub mod constraints;
-pub mod constraints64;
 pub mod tables;
-pub mod tables64;
 pub mod tests;
 pub mod utils;
