+++ prover/src/tables/bitwise.rs
