+++ prover/src/tables/lt.rs
