+++ prover/src/tests/bitwise_bus_tests.rs
@@ -19,7 +19,7 @@ use stark::trace::TraceTable;
 use stark::traits::AIR;
 use stark::verifier::{IsStarkVerifier, Verifier};
 
-use crate::tables64::types::{BusId, FE, GoldilocksExtension, GoldilocksField};
+use crate::tables::types::{BusId, FE, GoldilocksExtension, GoldilocksField};
 
 type F = GoldilocksField;
 type E = GoldilocksExtension;
