+++ prover/src/tests/bitwise_tests.rs
@@ -1,9 +1,9 @@
 //! Tests for the BITWISE precomputed table.
 
-use crate::tables64::bitwise::{
+use crate::tables::bitwise::{
     NUM_ROWS, bus_interactions, cols, generate_bitwise_trace, row_index,
 };
-use crate::tables64::types::FE;
+use crate::tables::types::FE;
 
 #[test]
 fn test_row_index() {
