+++ prover/src/tests/mod.rs
@@ -1,13 +1,9 @@
-pub mod cpu_table;
-pub mod integration_tests;
-
-// 64-bit VM tests
 #[cfg(test)]
 pub mod bitwise_bus_tests;
 #[cfg(test)]
 pub mod bitwise_tests;
 #[cfg(test)]
-pub mod constraints64_tests;
+pub mod constraints_tests;
 #[cfg(test)]
 pub mod cpu_tests;
 #[cfg(test)]
