+++ prover/src/tests/prove_elfs_tests.rs
@@ -25,19 +25,19 @@ use stark::trace::TraceTable;
 use stark::traits::AIR;
 use stark::verifier::{IsStarkVerifier, Verifier};
 
-use crate::constraints64::cpu::create_all_cpu_constraints;
-use crate::tables64::bitwise::{
+use crate::constraints::cpu::create_all_cpu_constraints;
+use crate::tables::bitwise::{
     BitwiseLookup, bus_interactions as bitwise_bus_interactions, cols as bitwise_cols,
     generate_bitwise_trace, update_multiplicities,
 };
-use crate::tables64::cpu::{
+use crate::tables::cpu::{
     bus_interactions as cpu_bus_interactions, collect_bitwise_lookups_from_logs,
     generate_cpu_trace_from_logs,
 };
-use crate::tables64::lt::{
+use crate::tables::lt::{
     LtOperation, bus_interactions as lt_bus_interactions, generate_lt_trace,
 };
-use crate::tables64::types::{GoldilocksExtension, GoldilocksField};
+use crate::tables::types::{GoldilocksExtension, GoldilocksField};
 
 type F = GoldilocksField;
 type E = GoldilocksExtension;
@@ -83,7 +83,7 @@ fn create_cpu_air(
     };
 
     AirWithBuses::new(
-        crate::tables64::cpu::cols::NUM_COLUMNS,
+        crate::tables::cpu::cols::NUM_COLUMNS,
         auxiliary_trace_build_data,
         proof_options,
         1,
@@ -101,7 +101,7 @@ fn create_bitwise_air(
     };
 
     AirWithBuses::new(
-        crate::tables64::bitwise::cols::NUM_COLUMNS,
+        crate::tables::bitwise::cols::NUM_COLUMNS,
         auxiliary_trace_build_data,
         proof_options,
         1,
@@ -119,7 +119,7 @@ fn create_lt_air(
     };
 
     AirWithBuses::new(
-        crate::tables64::lt::cols::NUM_COLUMNS,
+        crate::tables::lt::cols::NUM_COLUMNS,
         auxiliary_trace_build_data,
         proof_options,
         1,
@@ -387,7 +387,7 @@ fn test_cpu_only_no_bus() {
     };
     let cpu_air: AirWithBuses<F, E, stark::lookup::NullBoundaryConstraintBuilder, ()> =
         AirWithBuses::new(
-            crate::tables64::cpu::cols::NUM_COLUMNS,
+            crate::tables::cpu::cols::NUM_COLUMNS,
             auxiliary_trace_build_data,
             &proof_options,
             1,
